requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
click==8.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
isort==7.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pytokens==0.2.0
ruff==0.14.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.22.1
